requirements.txt
asgiref==3.11.1
asttokens==3.0.1
attrs==26.1.0
autobahn==25.12.2
automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.7
comm==0.2.3
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
debugpy==1.8.20
decorator==5.2.1
django==6.0.4
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
executing==2.2.1
gunicorn==25.3.0
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
ipykernel==7.2.0
ipython==9.12.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jupyter-client==8.8.0
jupyter-core==5.9.1
matplotlib-inline==0.2.1
msgpack==1.1.2
mysqlclient==2.2.8
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.6
prompt-toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pygments==2.20.0
pyjwt==2.12.1
pyopenssl==26.0.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.33.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
wcwidth==0.6.0
zope-interface==8.3
